Add Titans perf instrumentation

+++ src/main/java/net/minecraft/entity/titan/EntityGargoyle.java
@@ -132,9 +132,9 @@ extends EntityGolem {
     protected void entityInit() {
         super.entityInit();
         this.dataWatcher.addObject(13, (Object)Byte.valueOf((byte)0));
-        this.dataWatcher.addObject(16, (Object)Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(16, (Object)0);
         this.dataWatcher.addObject(17, (Object)0);
-        this.getDataWatcher().addObject(21, (Object)Byte.valueOf((byte)0));
+        this.getDataWatcher().addObject(21, (Object)0);
     }
 
     public int getGargoyleType() {
